add new test case (#144)
* add new test case

+++ tests/test_cli.py
@@ -2,10 +2,11 @@
 # Licensed under the MIT License.
 
 import os
-import platform
 import pytest
+import platform
 import re
 import subprocess
+import shutil
 import time
 from click.testing import CliRunner
 from dotenv import load_dotenv
@@ -20,8 +21,15 @@ if os.path.exists(filename):
     load_dotenv(filename)
 docker_client = EdgeDockerClient()
 
-VALID_DEVICECONNECTIONSTRING = os.environ[platform.system().upper() + '_DEVICE_CONNECTION_STRING']
+tests_dir = os.path.join(workingdirectory, "tests")
+test_config_dir = os.path.join(tests_dir, "assets", "config")
+test_resources_dir = os.path.join(tests_dir, 'test_compose_resources')
+
 VALID_IOTHUBCONNECTIONSTRING = os.environ['IOTHUB_CONNECTION_STRING']
+VALID_DEVICECONNECTIONSTRING = os.environ[platform.system().upper() + '_DEVICE_CONNECTION_STRING']
+VALID_CONTAINERREGISTRYSERVER = os.environ['CONTAINER_REGISTRY_SERVER']
+VALID_CONTAINERREGISTRYUSERNAME = os.environ['CONTAINER_REGISTRY_USERNAME']
+VALID_CONTAINERREGISTRYPASSWORD = os.environ['CONTAINER_REGISTRY_PASSWORD']
 
 device_id = re.findall(".*DeviceId=(.*);SharedAccessKey.*", VALID_DEVICECONNECTIONSTRING)[0]
 iothub_name = re.findall(".*HostName=(.*);DeviceId.*", VALID_DEVICECONNECTIONSTRING)[0].split('.')[0]
@@ -30,6 +38,38 @@ iothub_name = re.findall(".*HostName=(.*);DeviceId.*", VALID_DEVICECONNECTIONSTR
 @pytest.fixture
 def runner():
     return CliRunner()
+
+
+def get_docker_os_type():
+    return docker_client.get_os_type().lower()
+
+
+def docker_login(username, password, server):
+    start_process(['docker', 'login', '-u', username, '-p', password, server], False)
+
+
+def docker_logout(server_name):
+    start_process(['docker', 'logout', server_name], False)
+
+
+def docker_pull_image(image_name):
+    start_process(['docker', 'pull', image_name], False)
+
+
+def docker_tag_image(original_tag_name, new_tag_name):
+    start_process(['docker', 'tag', original_tag_name, new_tag_name], False)
+
+
+def docker_push_image(image_name):
+    start_process(['docker', 'push', image_name], False)
+
+
+def get_platform_type():
+    if get_docker_os_type() == 'windows':
+        platform_type = 'windows-amd64'
+    else:
+        platform_type = 'amd64'
+    return platform_type
 
 
 def update_setting_ini_as_firsttime():
@@ -49,16 +89,13 @@ def start_process(command, is_shell):
                                stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     output, error = process.communicate()
     if process.returncode == 0:
-        print('Process Output Message:\n' + str(output) + '\n')
+        return str(output)
     else:
         command = command.replace(VALID_IOTHUBCONNECTIONSTRING, '******')
         raise Exception(error)
-    return str(output)
 
 
-def cli_start_with_deployment(runner, deployment_file):
-    test_resources_dir = os.path.join('tests', 'test_compose_resources')
-    deployment_json_file_path = os.path.join(test_resources_dir, deployment_file)
+def cli_start_with_deployment(runner, deployment_json_file_path):
     result = runner.invoke(cli.start, ['-d', deployment_json_file_path])
     if 'IoT Edge Simulator has been started in solution mode' not in result.output.strip():
         raise Exception(result.stdout)
@@ -88,8 +125,8 @@ def cli_stop(runner):
         raise Exception(result.stdout)
 
 
-def verify_docker_output(docker_cmd, expect_values):
-    result = start_process(docker_cmd, False)
+def verify_docker_output(docker_cmd, expect_values, use_shell):
+    result = start_process(docker_cmd, use_shell)
     print('Process result is: \n %s \n' % (result))
     for expect_value in expect_values:
         if expect_value in result:
@@ -100,9 +137,9 @@ def verify_docker_output(docker_cmd, expect_values):
     return True
 
 
-def wait_verify_docker_output(docker_cmd, expect_values):
+def wait_verify_docker_output(docker_cmd, expect_values, use_shell=False):
     times = 0
-    while (not verify_docker_output(docker_cmd, expect_values)):
+    while (not verify_docker_output(docker_cmd, expect_values, use_shell)):
         print("Waiting until docker command is ready... ...\n")
         time.sleep(10)
         times += 1
@@ -110,19 +147,50 @@ def wait_verify_docker_output(docker_cmd, expect_values):
             raise Exception('Timeout to wait until it appears expected value ' + str(expect_values))
 
 
+def get_all_docker_volumes():
+    output = start_process(['docker', 'volume', 'ls'], False)
+    return output
+
+
 def remove_docker_volumes(volumes):
+    all_volumes = get_all_docker_volumes()
     for volume in volumes:
-        start_process(['docker', 'volume', 'rm', volume, '-f'], False)
+        if volume in all_volumes:
+            start_process(['docker', 'volume', 'rm', volume, '-f'], False)
+
+
+def get_all_docker_images():
+    output = start_process(['docker', 'image', 'ls'], False)
+    return output
 
 
 def remove_docker_images(images):
+    all_images = get_all_docker_images()
     for image in images:
-        start_process(['docker', 'rmi', '-f', image], False)
+        image_without_tag = image.split(':')[0]
+        if image_without_tag in all_images:
+            start_process(['docker', 'rmi', '-f', image], False)
+
+
+def get_all_docker_networks():
+    output = start_process(['docker', 'network', 'ls'], False)
+    return output
 
 
 def remove_docker_networks(networks):
+    all_networks = get_all_docker_networks()
     for network in networks:
-        start_process(['docker', 'network', 'rm', network], False)
+        if network in all_networks:
+            start_process(['docker', 'network', 'rm', network], False)
+
+
+def update_file_content(file_path, actual_value, expected_value):
+    with open(file_path, "r+") as f:
+        stream_data = f.read()
+        ret = re.sub(actual_value, expected_value, stream_data)
+        f.seek(0)
+        f.truncate()
+        f.write(ret)
 
 
 def test_cli(runner):
@@ -132,11 +200,12 @@ def test_cli(runner):
     assert 'Usage: main' in result.output.strip()
 
 
-@pytest.mark.skipif(docker_client.get_os_type().lower() == 'windows', reason='It does not support windows container')
+@pytest.mark.skipif(get_docker_os_type() == 'windows', reason='It does not support windows container')
 def test_cli_start_with_deployment(runner):
     try:
         cli_setup(runner)
-        cli_start_with_deployment(runner, 'deployment_without_custom_module.json')
+        deployment_json_file_path = os.path.join(test_resources_dir, 'deployment_without_custom_module.json')
+        cli_start_with_deployment(runner, deployment_json_file_path)
         wait_verify_docker_output(['docker', 'logs', 'edgeHubDev'], ['Opened link'])
         wait_verify_docker_output(['docker', 'logs', 'tempSensor'], ['Sending message'])
         invoke_module_method()
@@ -196,29 +265,51 @@ def test_cli_output_modulecred_file(runner):
             os.remove(output_file_path)
 
 
-@pytest.mark.skipif(docker_client.get_os_type().lower() == 'windows', reason='It does not support windows container')
+@pytest.mark.skipif(get_docker_os_type() == 'windows', reason='It does not support windows container')
 def test_cli_create_options_for_custom_volume(runner):
     try:
+        deployment_json_file_path = os.path.join(test_resources_dir, 'deployment_with_custom_volume.json')
+        temp_config_folder = os.path.join(tests_dir, 'config_tmp')
+        os.makedirs(temp_config_folder)
+
+        config_file_path = os.path.join(temp_config_folder, 'deployment_with_custom_volume.json')
+        shutil.copy(deployment_json_file_path, config_file_path)
+
+        if get_docker_os_type() == "windows":
+            update_file_content(config_file_path, '/mnt_test', 'C:/mnt_test')
+
         cli_setup(runner)
-        remove_docker_volumes(['testVolume'])
-        cli_start_with_deployment(runner, 'deployment_with_custom_volume.json')
-        wait_verify_docker_output(['docker', 'volume', 'ls'],
-                                  ['testVolume', 'edgemoduledev', 'edgehubdev'])
-        expected_volumes = (['"Source": "testVolume"',
-                             '"Destination": "/mnt_test"',
-                             '"Source": "edgemoduledev"',
-                             '"Destination": "/mnt/edgemodule"'])
-        wait_verify_docker_output(['docker', 'inspect', 'tempSensor'], expected_volumes)
-        wait_verify_docker_output(['docker', 'inspect', 'edgehubdev'],
-                                  '"Mountpoint": "/var/lib/docker/volumes/edgehubdev/_data"')
+        remove_docker_volumes(['testVolume', 'testvolume'])
+        cli_start_with_deployment(runner, config_file_path)
+
+        if get_docker_os_type() == 'linux':
+            expected_volumes = (['testVolume', 'edgemoduledev', 'edgehubdev'])
+            expected_tempsensor_volumes = (['"Source": "testVolume"',
+                                            '"Target": "/mnt_test"',
+                                            '"Source": "edgemoduledev"',
+                                            '"Target": "/mnt/edgemodule"'])
+            expected_edgehubdev_volumes = (['"Source": "edgehubdev"', '"Target": "/mnt/edgehub"'])
+        elif get_docker_os_type() == 'windows':
+            expected_volumes = (['testvolume', 'edgemoduledev', 'edgehubdev'])
+            expected_tempsensor_volumes = (['"Source": "testVolume"',
+                                            '"Target": "C:/mnt_test"',
+                                            '"Source": "edgemoduledev"',
+                                            '"Target": "c:/mnt/edgemodule"'])
+            expected_edgehubdev_volumes = (['"Source": "edgehubdev"',
+                                            '"Target": "c:/mnt/edgehub"'])
+
+        wait_verify_docker_output(['docker', 'volume', 'ls'], expected_volumes)
+        wait_verify_docker_output(['docker', 'inspect', 'tempSensor'], expected_tempsensor_volumes)
+        wait_verify_docker_output(['docker', 'inspect', 'edgeHubDev'], expected_edgehubdev_volumes)
     finally:
+        shutil.rmtree(temp_config_folder, ignore_errors=True)
         result = cli_stop(runner)
         assert 'IoT Edge Simulator has been stopped successfully' in result.output.strip()
         remove_docker_networks(['azure-iot-edge-dev'])
         remove_docker_images(['mcr.microsoft.com/azureiotedge-simulated-temperature-sensor:1.0',
                               'mcr.microsoft.com/azureiotedge-hub:1.0',
                               'hello-world'])
-        remove_docker_volumes(['testVolume'])
+        remove_docker_volumes(['testVolume', 'testvolume'])
 
 
 def test_cli_modulecred_for_multiple_modules(runner):
@@ -229,11 +320,12 @@ def test_cli_modulecred_for_multiple_modules(runner):
     assert 'ModuleId=edgeHubDev' in result.output.strip()
 
 
-@pytest.mark.skipif(docker_client.get_os_type().lower() == 'windows', reason='It does not support windows container')
+@pytest.mark.skipif(get_docker_os_type() == 'windows', reason='It does not support windows container')
 def test_cli_start_with_chunked_create_options(runner):
     try:
         cli_setup(runner)
-        cli_start_with_deployment(runner, 'deployment_with_create_options.json')
+        deployment_json_file_path = os.path.join(test_resources_dir, 'deployment_with_create_options.json')
+        cli_start_with_deployment(runner, deployment_json_file_path)
         wait_verify_docker_output(['docker', 'logs', 'edgeHubDev'], ['Opened link'])
         wait_verify_docker_output(['docker', 'logs', 'tempSensor'], ['Sending message'])
         expect_createoptions = ['FOO1=bar1', 'FOO2=bar2', 'FOO3=bar3',
@@ -248,19 +340,108 @@ def test_cli_start_with_chunked_create_options(runner):
                               'hello-world'])
 
 
+@pytest.mark.skipif(get_docker_os_type() == 'windows', reason='The base image of edgeHubDev image is 1809 but agent is 1803.'
+                    'So it does not support windows container.')
 def test_cli_start_with_input(runner):
     try:
         result = runner.invoke(cli.start, ['-i', 'input1'])
-        assert not result.exception
-        assert result.exit_code == 0
-        assert 'IoT Edge Simulator has been started in single module mode' in result.output.strip()
-        assert 'curl --header' in result.output.strip()
+        output = result.output.strip()
+        if result.exit_code == 0:
+            assert 'IoT Edge Simulator has been started in single module mode' in output
+            assert 'curl --header' in output
+        else:
+            raise Exception(result.stdout)
     finally:
         result = cli_stop(runner)
         assert 'IoT Edge Simulator has been stopped successfully' in result.output.strip()
         remove_docker_networks(['azure-iot-edge-dev'])
         remove_docker_images(['mcr.microsoft.com/azureiotedge-hub:1.0',
                               'mcr.microsoft.com/azureiotedge-testing-utility:1.0.0-rc1'])
+
+
+@pytest.mark.skipif(get_docker_os_type() == 'windows', reason='The base image of edgeHubDev image is 1809 but agent is 1803.'
+                    'So it does not support windows container.')
+def test_cli_start_with_create_options_for_bind(runner):
+    try:
+        deployment_json_file_path = os.path.join(test_resources_dir, 'deployment_with_create_options_for_bind.json')
+        temp_config_folder = os.path.join(tests_dir, 'tmp_config')
+        os.makedirs(temp_config_folder)
+
+        config_file_path = os.path.join(temp_config_folder, 'deployment_with_create_options_for_bind.json')
+        shutil.copy(deployment_json_file_path, config_file_path)
+
+        if get_docker_os_type() == "windows":
+            update_file_content(config_file_path, '/usr:/home/moduleuser/test',
+                                r'C:\\\\\\\\Users:C:/moduleuser/test')
+
+        cli_setup(runner)
+        cli_start_with_deployment(runner, config_file_path)
+
+        wait_verify_docker_output(['docker', 'logs', 'edgeHubDev'], ['Opened link'])
+        wait_verify_docker_output(['docker', 'logs', 'tempSensor'], ['Sending message'])
+        if get_docker_os_type() == "windows":
+            wait_verify_docker_output('echo dir | docker exec -i -w c:/moduleuser/test/ tempSensor cmd', ['Public'], True)
+        else:
+            wait_verify_docker_output(['docker', 'exec', 'tempSensor', 'ls', '/home/moduleuser/test'], ["share"])
+    finally:
+        shutil.rmtree(temp_config_folder, ignore_errors=True)
+        result = cli_stop(runner)
+        assert 'IoT Edge Simulator has been stopped successfully' in result.output.strip()
+        remove_docker_networks(['azure-iot-edge-dev'])
+        remove_docker_images(['mcr.microsoft.com/azureiotedge-simulated-temperature-sensor:1.0',
+                              'mcr.microsoft.com/azureiotedge-hub:1.0',
+                              'hello-world'])
+
+
+@pytest.mark.skipif(get_docker_os_type() == 'windows', reason='The base image of edgeHubDev image is 1809 but agent is 1803.'
+                    'So it does not support windows container.')
+def test_cli_start_with_registry(runner):
+    temp_config_folder = os.path.join(tests_dir, 'deployment_config')
+    os.makedirs(temp_config_folder)
+    try:
+        old_tag = '1.0.0-rc1'
+        new_tag = '1.0.0-rc1.test'
+        image_name = 'azureiotedge-testing-utility'
+        old_image_name = image_name + ':' + old_tag
+        new_image_name = image_name + ':' + new_tag
+        old_tag_name = ('{0}/{1}').format('mcr.microsoft.com', old_image_name)
+        new_tag_name = ('{0}/{1}').format(VALID_CONTAINERREGISTRYSERVER, new_image_name)
+
+        docker_pull_image(old_tag_name)
+        docker_tag_image(old_tag_name, new_tag_name)
+        docker_login(VALID_CONTAINERREGISTRYUSERNAME, VALID_CONTAINERREGISTRYPASSWORD, VALID_CONTAINERREGISTRYSERVER)
+        docker_push_image(new_tag_name)
+
+        config_file_path = os.path.join(temp_config_folder, 'deployment.json')
+        deployment_file_path = os.path.join(test_config_dir, "deployment.json")
+        shutil.copy(deployment_file_path, config_file_path)
+
+        module_image_name = ('"image": "{0}/{1}"').format(VALID_CONTAINERREGISTRYSERVER, new_image_name)
+        update_file_content(config_file_path, '"image": ""', module_image_name)
+
+        update_file_content(config_file_path, '"username": ""',
+                            '"username": "' + VALID_CONTAINERREGISTRYUSERNAME + '"')
+        update_file_content(config_file_path, '"password": ""',
+                            '"password": "' + VALID_CONTAINERREGISTRYPASSWORD + '"')
+        update_file_content(config_file_path, '"address": ""',
+                            '"address": "' + VALID_CONTAINERREGISTRYSERVER + '"')
+
+        remove_docker_images([old_tag_name, new_tag_name])
+        docker_logout(VALID_CONTAINERREGISTRYSERVER)
+
+        cli_start_with_deployment(runner, config_file_path)
+        wait_verify_docker_output(['docker', 'ps'], ['azureiotedge-testing-utility', new_tag])
+    finally:
+        shutil.rmtree(temp_config_folder, ignore_errors=True)
+
+        result = cli_stop(runner)
+        assert 'IoT Edge Simulator has been stopped successfully' in result.output.strip()
+
+        remove_docker_networks(['azure-iot-edge-dev'])
+        remove_docker_images([new_tag_name,
+                              old_tag_name,
+                              'mcr.microsoft.com/azureiotedge-hub:1.0',
+                              'hello-world'])
 
 
 # def test_cli_with_option(runner):
